python/requirements.txt
Flask
prometheus_client
pyjks
pyOpenSSL